before union with kolya

+++ Printf/includes/conversions.h
@@ -6,7 +6,7 @@
 /*   By: artemstarshov <artemstarshov@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/04 21:41:55 by artemstarsh       #+#    #+#             */
-/*   Updated: 2020/01/14 03:17:38 by artemstarsh      ###   ########.fr       */
+/*   Updated: 2020/01/16 01:55:25 by artemstarsh      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,10 @@
 # define ALL_CONVERSIONS_H
 
 # include "flags_describer.h"
+# include <limits.h>
 
 typedef long long	t_signed;
-typedef size_t		t_unsigned;
+typedef uintmax_t	t_unsigned;
 typedef long double	t_float;
 
 char	*signed_conversion(va_list args_list, const flags_describer *describer);
